refactor : admin 패키지 삭제
admin 에서 User(seller, customer) 의 인증 관련 코드는
user 패키지로 이동

admin 에서 인증메일 관련 코드는 새로운 mail 패키지를 만들어서
이곳으로 이동
