+++ src/main/java/com/itonse/clotheslink/mail/controller/MailController.java
@@ -1,7 +1,6 @@
-package com.itonse.clotheslink.admin.controller;
+package com.itonse.clotheslink.mail.controller;
 
-import com.itonse.clotheslink.admin.service.MailAuthService;
-import com.itonse.clotheslink.admin.service.TokenService;
+import com.itonse.clotheslink.mail.service.MailAuthService;
 import com.itonse.clotheslink.common.UserVo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -9,17 +8,10 @@ import org.springframework.web.bind.annotation.*;
 
 @RequiredArgsConstructor
 @RestController
-public class AdminController {
+public class MailController {
 
-    private final TokenService tokenService;
     private final MailAuthService mailAuthService;
 
-    @GetMapping("/token/validation")
-    public ResponseEntity<UserVo> validateTokenTest(
-            @RequestHeader(name = "Authorization") String token) {
-
-        return ResponseEntity.ok().body(tokenService.validateToken(token));
-    }
 
     @PostMapping("/auth/mail")
     public ResponseEntity<UserVo> sendAuthMail(@RequestHeader(name = "Authorization") String token) {
