+++ src/main/java/com/itonse/clotheslink/user/service/CustomerAuthService.java
@@ -1,5 +1,6 @@
 package com.itonse.clotheslink.user.service;
 
+import com.itonse.clotheslink.user.domain.Customer;
 import com.itonse.clotheslink.user.dto.SignInDto;
 import com.itonse.clotheslink.user.dto.SignUpDto;
 import com.itonse.clotheslink.user.dto.UserInfoResponse;
@@ -9,4 +10,6 @@ public interface CustomerAuthService {
     UserInfoResponse signUp(SignUpDto dto);
 
     String signIn(SignInDto dto);
+
+    Customer findCustomerByToken(String token);
 }
