+++ src/main/java/com/itonse/clotheslink/user/service/SellerAuthService.java
@@ -1,5 +1,6 @@
 package com.itonse.clotheslink.user.service;
 
+import com.itonse.clotheslink.user.domain.Seller;
 import com.itonse.clotheslink.user.dto.SignInDto;
 import com.itonse.clotheslink.user.dto.UserInfoResponse;
 import com.itonse.clotheslink.user.dto.SignUpDto;
@@ -8,4 +9,7 @@ public interface SellerAuthService {
     UserInfoResponse signUp(SignUpDto dto);
 
     String signin(SignInDto dto);
+
+    Seller findSellerByToken(String token);
+
 }
